requirements_3.txt
python-dateutil~=2.8.2
multidict~=4.7.6
pyyaml~=5.4.1
jsonschema~=3.2.0
PyGithub~=1.55
pathlib~=1.0.1
py~=1.10.0
colorama~=0.4.4
pip>=21.0.1
argparse~=1.4.0
cryptography~=3.4.7
lxml~=4.6.3
attrs~=21.2.0
Werkzeug~=2.0.1
itsdangerous~=2.0.1
MarkupSafe~=2.0.1
six~=1.16.0
pluggy~=0.13.1
packaging~=21.0
atomicwrites~=1.4.0
certifi~=2021.5.30
chardet~=3.0.4
urllib3~=1.26.6
toml~=0.10.2
idna~=3.2
greenlet~=1.1.0
clickclick~=20.10.2
decorator~=5.0.9
inflection~=0.5.1
pyparsing~=2.4.7
pyrsistent~=0.18.0
rfc3987~=1.3.8
setuptools~=54.2.0
docutils~=0.16
validators~=0.18.2
smart_open>=5.1.0
pytest~=6.2.4
pytest-asyncio
PyJWT
pynacl
pyld